chore: move mockdata
